services/db-sink/requirements.txt
# DB Sink Service - Requirements
confluent-kafka>=2.3.0
psycopg2-binary>=2.9.9
